Streamlit App
Added Streamlit code for ML model

This commit adds Streamlit code to the app.py file that can be used to deploy the ML model to a web app. The code uses Streamlit to create a user interface that allows users to input data and get predictions from the model.

The code was tested on a sample dataset and was found to work correctly.